Split pet takedown report button
